Added XPath predicates, abbreviated context & parent node, better context flow.
The execution flow for XPath querying now passes contexts around instead of
the document, leading to a more harmonious design overall. Absolute path
semantics are accomplished by using any node in context to crawl back up to
the document (the soup) and traversal begins with a context containing only
the document.

The nud and led functions now take an "expression_fn," that being the
"expression" function from the parsing module. This is needed for constructs,
such as the beginning of a predicate, where a sub-expression needs to be
parsed forward.

Also changed node type function names so they look more graceful when used in
functional contexts.