Functions & binary ops (sort of); node set ==; unified output; parse refactor.
There's now very limited support for XPath function calls (one-argument and
one argument exactly) and binary operators (plus sign for numeric addition),
which required some further jamming of this parsing algorithm into my thick
skull. Thank you, Doug Crockford!

    http://javascript.crockford.com/tdop/tdop.html

XPath functions are now dynamically loaded from modules in the
"hq.xpath.functions" package, and each such module has to define a global
"exports" list with the (local) names of classes and functions to be used as
XPath functions.

I started approaching the filling-in of equality testing for XPath objects
based on a top-down reading of the standard (1.0), leading me to start with
node set equality. This in turn led me to construction of string values from
nodes. Equality testing is still a work in progress, however, as one can see
from the NotImplementedErrors and a skipped test in the code.

I spruced up the XPath parsing module and its relationship to tokens because I
was having to make changes here to facilitate function calls (advancing the
parser), and it was time to apply lessons learned and attack blight. I renamed
the global "next," by the way, so that I could (straightforwardly) use the
Python built-in "next."

The means of output had diverged between the tests and the CLI, and the CLI's
ideas were naive about XPath object types (considering nodes only), so I fixed
all of that up.

Speaking of tests, they caught my lack of bool coercion behavior for Python
2.x, leading to the __nonzero__ on the boolean XPath object class.

The work here is test-driven, strictly speaking, but many more tests needed.