Finished arithmetic ops; token interpretation branching; tested n-ary functions
The "div" and "mod" operators, when I introduced them as tokens, broke all of
the tests using "div" as a node name test, so it's a good thing (a) I wrote
tests and (b) HTML has an element called "div." I made an adaptation that
conforms, I believe, with the design of the parsing algorithm, retaining the
last token so that branching logic can base decisions on this small context
clue.

I took advantage of the need to test this token-branching behavior as an
occasion to test function calls of more than one argument, introducing a
couple of core string functions that take multiple arguments. This led to the
hyphens-to-underscores replacement to retain correspondence between function &
class names in the source and function names in XPath.

Having added the wrap_body option to my process_xpath_query test utility
function, I should go back and simplify some other tests with it, but it
didn't happen in this check-in.