Abbreviated node() test (*); test already implemented axes; renamed tokens.
I realized, as I was starting to fill in tests for axes that are already
implemented, that I hadn't implemented the asterisk yet, and it was about
time.

I also got tripped up a little by the rude intrusion of "unicode" as a string
type in 2.7. Despite appearances, I was actually vaguely aware of this factor
but chose (and continue choosing) to defer proper and complete consideration
of it.