Merged XPath NameTest and NodeTest classes (finally).
The try/except in NodeTest.apply_to_child is needed for the case when a test
is applied to the document root that doesn't match anything. NameTest worked
differently, and the interpretation_of_div_and_mod_as_operators_or_name_tests
test caught it (fourth assertion). I had also failed to move the node set
reversal to preserve document order in one spot in NodeTest (it was working,
and only had a test covering it, in NameTest).