Implemented abbreviated attribute axis (incl. @*) & fixed wrong // evaluation.
I realized the unsustainability of my cheap "descendant axis" trick in double-
slash evaluation when I tried to follow it up with the attribute axis. So now
"//" should work as it's supposed to (I think).

I also made some import changes to fix import errors running under Python 2.7
and I beefed up verbose output from the token classes a bit. And I fixed a
couple of issues related to setup.py because I noticed it was broken when I
tried to run it (to run "clean," when I was first seeing import problems).