New, broader solution for document-order results.
All of the reversing wasn't cutting it, as I found after writing a test to
flesh out the double-slash mid-path and found that evaluation of the child
axis during the // traversal led unavoidably to non-document order in a way
that was difficult to patch up with the same sort of "reverse" kludge. So now
I'm marking every node in the tree with an incrementing "hq_doc_index" value
representing document order, and also an "hq_attrs" list with a pre-
constructed, pre-sorted list of AttributeNodes.

A few tests were expecting non-document-order results, and after checking
against xmllint (which always returns stuff in document order), I changed the
tests.