Simplified NodeTest; reverse-order node sets for preceding axes
For the position() function to work in predicates that follow preceding and
preceding-sibling paths, the node sets needed to be in reverse document order.
In the course of exploring this issue, I realized that the algorithmic design
in the NodeTest class wasn't a good match for the actual XPath semantics, so I
re-factored it in a way that simplifies the class and hopefully doesn't do too
much harm to performance.

I also fixed the handling of argument flags in hq.py, as I realized this stuff
wasn't working. There's also some verbose output housekeeping in tokens.py, as
usual.