Feel-dumb check-in of the year: paths now evaluate recursively left-to-right.
I remember puzzling, mouth agape, about how a location path should be modeled
by the parser. Typically, I made an ignorant and wrong-headed choice, having
the parser treat each step like a sub-expression and chaining evaluation
right-to-left, accumulating results at each step and iterating contexts across
the entire accumulated result set. This is not how the spec says it works, of
course, and I realized this as I was tinkering with the software, coming up
with a demo for my son. The problem I encountered is embodied in the two new
test in this check-in, which used to fail because the value of the position()
function was relative to the set of ALL leaf elements in the path so far (1
through 4), rather than being relative to each group generated from the
penultimate step or predicate as one descends recursively (1 and 2, then 1 and
2).

With this parsing method, if you want to pin evaluation to the left-hand end
of a sub-expression, you need to hoover the expression from the left, so now
there are parsing functions dedicated to location paths in query_xpath.py, and
all of the token classes associated with paths are vastly simplified.
Evaluation functions in general no longer accept an optional axis argument, as
that was a subsidiary kludge from the way tokens used to evaluate paths.