Started Unicode work; dev vs. release requirements.txt; hyphens in name tests.
Proper Unicode support under Python 2.7 is going to be an ongoing project, but
at least you can feed it HTML with Latin 1 characters in it, now, and it won't
raise exceptions as it's trying to print the results. I made the tests for
this area operate on the actual CLI, rather than calling the query_xpath
function directly, because I wanted to verify the entire input/output stack.

I wasn't allowing hyphens in name tests (elements or attributes), as it turns
out, so I added tests and fixed the XPath tokenizer.

I broke up requirements so that the dev-only stuff won't get shipped with the
product (when I eventually get around to uploading it). And speaking of
breaking stuff up, I split up the test_utils because only some of them are
XPath-specific. Plus, I moved some tests from the "name test tests" module to
the "axis tests" module because they really belonged there (this was early
stuff).

I also stopped printing my command-line arguments, since I now have tests that
operate on the CLI, and it's about time, anyway.