CLI tests (options, output) + dropped CSS + text normalize & --preserve flag.
Text content (at least element content, for the moment) is now automatically
normalized when its string value is derived. This is a sensical and convenient
default that will produce much terser queries than those filled with explicit
normalize-space calls.

I also added the --preserve flag to compensate for this feature. I believe it
makes the most sense for this kind of state to be part of the "dynamic
context," as the XQuery standard describes it, so that it can be changed, for
example, in embedded parsing contexts (such as interpolated strings). Because
of the "evaluate in context" functions' existing dependency on make_node_set,
coming from the object_type module that now needs visibility into context (in
the string_value function), I had to break those evaluation functions out in a
separate module to avoid circular dependencies. Least of the evils, I thought.

When I started, I envisioned a tool that could do XPath plus as well as CSS
plus. Now I think that's overkill, and I want the tool to be about XPath plus
XQuery plus et ceteral, a.k.a., HQuery. So no more CSS flag in the CLI.

The CLI testing drove some cleanup elsewhere, like error messaging (which
touched the function_support module).