Added "filters" to interpolated string expressions, including chaining.
Filters are transforms you can apply to an escaped expression in an
interpolated string, doing things like joining the string values in a node set
with some delimiter in between, or truncating a string value to a certain
length with a suffix, like an ellipsis. They come before the expression
itself, just inside the open curly brace, and they take the form of a very
brief name (one letter, so far) followed by a colon-delimited list of
arguments (no optional arguments) with a trailing colon separating the filter
from the expression, or from the next filter if you are chaining filters.

For example:

   `The headings are: ${j:, ://h1 | //h2}`

Here, the "j" for "join" filter will essentially do the equivalent of
", ".join(node_set) in Python, where node_set is actually the collection of
string values from the node set generated by the XPath expression.

There's also a truncate ("t") filter with two arguments, length and suffix,
and the tests demonstrate both filter's use as well as the chaining feature.