Sequence construction and diverse numeric types (float and int).
Sequence construction required commas to have a non-nothing left-binding
power. This produced a problem for parsing functions with multiple parameters,
so I squared the circle by putting sequence construction at the bottom of the
priority ladder (which I suspect is correct) and making function parsing call
the expression function for each argument with the sequence LBP instead of
nothing. I added an assertion just to emphasize the need for this special LBP
to be next-to-nothing.

Now numeric behavior is closer to XQuery semantics rather than XPath 1.0.