Implemented rudimentary FLWOR + more functions for use case experiment.
The only FLWOR clauses supported so far are "let" and "return," and the tests
are still pretty anemic.

The function stuff I added was all motivated by a sample use case I'm toying
with, that being the extraction and reporting of information about EC2
instances from the table of instances in the AWS dashboard. The "matches"
function is based on the XPath/XQuery 3 function standard, but with liberties
taken to facilitate context-node-based usage in predicates, which makes for
more terse queries.

I figure variables work like a stack, similar to contexts but separate.

Ran across more Unicode issues in verbose printing from equality operator
functions.

It turns out, I guess, that docopts only passes along the long form of flags
if there are multiple forms (which the doc kind of says in one place, but it
follows that up with an example that appears to indicate otherwise, so I was
unsure). Thus the "--ugly" change.

Also added a few tests here and there to cover things I'd previously
neglected.