Implemented "for" iteration in FLWORs (only in-iteration LETs allowed).
FLWORs can now have a "for" clause, and iteration will happen over the
sequence in the normal way. The sequence doesn't need to be a node set. At
this point, however, all of the "let" declarations are treated as if they
follow the "for," in terms of XQuery-like semantics.

"Rooted" location paths are those that begin with a slash preceded by an
expression that produces a sequence, such as a variable reference (which is
treated the same as the sequence produced in the prior location path, from a
path evaluation perspective). These were needed to base location paths in
FLWOR return clauses on variables such as the iteration variable.

Variable references broke embedded expressions inside interpolated strings, so
I needed to make the algorithm that chunks up the interpolated string smarter.

I added sequences to the type "system" in ways that I had heretofore avoided,
on Simplicity Principle grounds, as these measures were needed to support the
general use of sequences as the basis of "for" iteration. I *think* I'm safe
treating sequence as "node set minus the is_any_tag constraint," until I see a
need for further specialization.

I also initiated a practice of consuming reserved words (and, later, arbitrary
names such as those one finds in XQuery computed constructors) as
NameTestTokens internally, those being tokens that are always advanced over
manually by specialized parser algorithms like FLWOR.

Plus, there's some verbosity housekeeping.