Abbreviated FLWORs + class() function + HTML escapes + interpolation fix.
These changes were the last necessary to implement a motivating use case that
I had sketched a while ago: producing a CSV file of episodes based on the
Wikipedia "Breaking Bad (Season 1)" page, including truncated episode
descriptions that require following-sibling:: to reach. Here's the HQuery that
I ended up with, broken up into multiple lines for readability:

   concat(
      `"${j:","://table[class("wikiepisodetable")]/tr[1]/th}","Plot"&#10;`,
      //table[class("wikiepisodetable")]/tr[position()mod2=0] ->
         `"${j:",":$_/*}","${t:50:...:$_/following-sibling::tr[1]/td}"&#10;`
   )

This required abbreviated FLWORs (the "->"), the class() function to make
selection of nodes based on their class attributes simple, and support for
HTML escapes (like "&#10;") in string literals. It also required that the
string value of a sequence (e.g., "string(<sequence value>)") take the form of
a join (concatenation) of the string values of all of the items in the
sequence. In regular XQuery, it's an error to pass a non-node-set,
non-singleton sequence to the string() function. But this behavior contributes
significantly to the brevity of that HQuery up there, as string_value is
implicitly applied by the concat() function to its second argument (a sequence
of strings).

Interpolated string evaluation was broken because of the use of reassigned
local variables in closures, so I fixed that.

I also did some more verbosity housekeeping; a habit, nowadays.