JSON hash constructor with optional filters.
There's now a JSON hash computed constructor that produces a JSON hash as a
string. It's very similar to the HTML element computed constructor, obviously,
but it interprets content differently (elements become attributes with the
tag name becoming the attribute name and the string value of the tag becoming
the value). This kind of constructor also supports an optional filter clause
that looks like the filters I designed for interpolated strings. For example:

   hash {a:p,h1:n:div,h1:} { content... }

This clause specifies two filters, each of which takes a single filter
argument that's a comma-delimited list of tag names. The "a" for "array"
filter causes element content whose tag name is found in the argument list to
have its value take the form of an array (instead of a string) even when there
aren't multiple instances of that tag name among the hash's contents. Multiple
like-named tags in the contents are automatically put into an array value
under the attribute name corresponding to the tag name, but this filter
ensures that a given tag name is treated this way even when there aren't
multiple instances of that tag name in the contents.

The other filter, "n" for "number," converts all values from element contents
whose tag names are found in the argument list to numeric values in the JSON
output (as opposed to strings).