BREVITY BASH! New axis abbreviations and CSS "class" axis.
Now there are symbolic abbreviations, all followed by the double-colon, for
the other axes that need them (like "@" for attribute, but with the addition
of the double-colons to make the parsing simple and to distinguish them from
standard XPath abbreviations). Look at the _abbreviations collection in
axis.py for reference.

There's also a new "class" axis (abbreviated ".::") that has the same
semantics as the "child" axis except the name test that follows it is compared
against the class names in each tag's "class" attribute (if present) instead
of its tag name.

I had to move axis token recognition up in front of abbreviated parent (".")
so the latter wouldn't get in the way of recognizing ".::".