Implemented XQuery if/then/else + adjusted advance_over_name semantics.
Sorry for the dumb, long name in hquery_processor.py, but I felt the reason
for deviating from the advance_over_name/NameTestToken tactic deserved some
kind of additional explanation in code, and I always feel a little more
confident about the future prospects with this documentation technique than
just a dead comment.

As for advance_over_name, I noticed that it's always used in contexts where a
name is required. So I changed the semantics from optional to expected.