JSON arrays & attrs; better text strip; output fixes; regex replace filter.
Most of this work was motivated by the documentation I was writing, as I
discovered shortcomings that needed to be addressed in order for the doc to
tell a better story.

The JSON hash and array types deserved their own classes, since arrays would
have been yet another case of "list" and also to be consistent.

Now the automatic space normalization that is applied to HTML-text-derived
strings is less buggy and includes substitution of regular spaces for non-
breaking spaces. This is all in service of making it easy by default to use
HTML-embedded text data in non-HTML (programmatic) contexts.

The even() and odd() functions are equivalent to "position() mod 2 = 0" and,
well, you get the idea. Matching syntax sugar with the "even" and "odd"
keywords for CSS' "nth-of-type."

The new clause added to parse_location_path() is there to handle the case
where the predicate's left brace is the first location path syntax we
encounter in a location path. This happens when you use a predicate on a
variable reference, like this:

   $var[predicate()]

Along with a new regex replace filter (3 parameters: search pattern,
replacement pattern, flags), the interpolated strings implementation now does
a better (and way less stupid) job of parsing I.S. contents. It also avoids
replacing HTML character escapes too early, as it was doing.

The improvements in output algorithms, as well as debug_dump_anything, are
partial but (obviously) incomplete steps towards cleaning up output and, more
generally, value-type-arbitrated behavior. Should I surrender and wrap some
classes around the remaining "raw" stuff, the Beautiful Soup classes? We'll
see.

There are several non-headline improvements here and there, like implementing
some more of the core function library, that a pretty self-explanatory.

I also adopted tox for local multi-virtualenv testing, though I'm not using it
in Travis (since their multi-virtualenv thing is so nice).

I made the change in how tests capture stdin and stderr, by the way, because I
wanted to have access to stderr (verbose) output when executing individual
tests that were only looking at stdout. I wanted a variable in the test's
context that I could look at when I needed to.

The hq_runner.py change, by the way, was just "chmod +x hq_runner.py".