rr filter on sequences; whitespace fixed across element boundaries; more docs
The regex replace filter now works on sequences as well as unitary objects,
emitting a sequence derived from the string values of the original sequence
items (with the replacement applied, of course).

My last fix to handle space padding between text runs that cross element
boundaries didn't quite fix enough, as it was adding spaces between text runs
even when there wasn't any space on either side of the tag in the original
HTML. The enhanced algorithm is ugly-looking but correct, at least.

I also renamed the truncate filter "tru:::" instead of "t:::" and fixed a bug
in string_value when it operates on an empty node set.