New README (post-wiki); read HTML from file (--file option).
I sped things up slightly by using the proper data structure for ID uniqueness checking in make_node_set.

I'm trying to be strict, now, about always using the longer version of docopts flags that have a long and a short version, even though I'm not entirely sure that this is what I'm supposed to do (the doc seems to say so).