Implemented switch-like "union decomposition" feature, using "=>" operator.
I considered typeswitch from XQuery, then thought of another way...

This operator requires a union on its left side and something that LOOKS LIKE
a union on its right side. Each node produced by the union is iterated over
and is available in the special "$_" variable on the right-hand side, just
like the abbreviated FLWOR, but the behavior of the faux union on the right-
hand side is different. The operator expects the same number of clauses (i.e.,
pipes) on its left side and its right, and for each node that it iterates
over, it evaluates the clause on the right side at the same position as the
clause on the left that first produced the node.

If more than one clause in the source union produce the same node, the clause
that that node corresponds to is the first one (left-to-right) that produced
it.